YFT SUPPLEMENTAL DATASET: Fully spatial simulated datasets (Simulation #: 1-33)

This R data file contains 1-33 simulated data sets at the 221 5x5 gird cell level, which matches the spatial resolution used in the OM. These data are provided for your information and in a usable format if participants wish to develop spatially structured models beyond the one and four area aggregations already provided.

data('YFT_221cell_observations_1_33_v2')

Format

A data frame 'YFT_221cell_observations_1_33_v2' with variables:

sim_1

Simulated aggregated 1 area datasets for YFT; Simulation #1-33

biol_dat

Biological data for YFT

The key lists in sim_X$ (e.g., sim_1$) include:

Sim_X : Simulated spatial datasets, X: sim # 1-100

obs: Simulated observations

simulated_ll_jpn_cpue_YYY: CPUE data for pseudo-year YYY

simulated_XXX_lf_YYY: Length frequencies for pseudo-year YYY

tagrel: Tag releases

***processtag_XXX***: Tag release in a given pseudo-year

tagrecs: Simulated tag recapture data

***observationtag_recapture_XXX_in_YYY***: tag recapture in a given pseudo-year from release cohort XXX

layer : Spatial grids defined within 'layer'

base: 1 = on the water, 0 = on land

cell: unique cell row and column numbers

constant: 1

latitude: latitude for each cell center

longitude: longitude for each cell center

region: IOTC YFT regions (R1a, R1b, R2, R3, R4, R5)

sst: Sea Surface Temperature (see Dunn et al. 2020)

clo: Chlorophyll (see Dunn et al. 2020)

fishing_ff_YY: ff: fishery = ps, trol, bb, gill, ll, other, hand, YY = pseudo-year

The lists in biol_dat include:

M: age varying natural mortality

age: first age to last age (pseudo-years)

maturity: maturity at age ogive (pseudo-years)

k: age varying growth coefficients

L: length (cm)

Linf: length infinity (cm)

Lmin: length minimum (cm)

a: length-weight scaling coefficient

b: length-weight shape parameter

LenAge_cv: length-age coefficient of variability

Examples

data('YFT_221cell_observations_1_33_v2')
#> Warning: data set 'YFT_221cell_observations_1_33_v2' not found
head(sim_1$obs$simulated_cpue_ll_jpn_86$data$obs)
#> Error in head(sim_1$obs$simulated_cpue_ll_jpn_86$data$obs): object 'sim_1' not found