YFT SUPPLEMENTAL DATASET: Fully spatial simulated datasets (Simulation #: 67-100)

This R data file contains 67-100 simulated data sets at the 221 5x5 grid cell level, which matches the spatial resolution used in the OM. These data are provided for your information and in a usable format if participants wish to develop spatially structured models beyond the one and four area aggregations already provided.

data('YFT_221cell_observations_67_100_v2')

Format

A data frame 'YFT_221cell_observations_67_100_v2' with variables:

sim_1

Simulated aggregated 1 area datasets for YFT; Simulation #67-100

biol_dat

Biological data for YFT

The key lists in sim_X$ (e.g., sim_1$) include:

Sim_X : Simulated spatial datasets, X: sim # 1-100

obs: Simulated observations

simulated_ll_jpn_cpue_YYY: CPUE data for pseudo-year YYY

simulated_XXX_lf_YYY: Length frequencies for pseudo-year YYY

tagrel: Tag releases

***processtag_XXX***: Tag release in a given pseudo-year

tagrecs: Simulated tag recapture data

***observationtag_recapture_XXX_in_YYY***: tag recapture in a given pseudo-year from release cohort XXX

layer : Spatial grids defined within 'layer' (including cells, latitude, longitude for each grid cell, and catch data for each fishery)

base: 1 = on the water, 0 = on land

cell: unique cell row and column numbers

constant: 1

latitude: latitude for each cell center

longitude: longitude for each cell center

region: IOTC YFT regions (R1a, R1b, R2, R3, R4, R5)

sst: Sea Surface Temperature (see Dunn et al. 2020)

clo: Chlorophyll (see Dunn et al. 2020)

fishing_ff_YY: ff: THIS IS THE CATCH DATA: fishery = ps, trol, bb, gill, ll, other, hand, YY = pseudo-year

The lists in biol_dat include:

M: age varying natural mortality

age: first age to last age (pseudo-years)

maturity: maturity at age ogive (pseudo-years)

k: age varying growth coefficients

L: length (cm)

Linf: length infinity (cm)

Lmin: length minimum (cm)

a: length-weight scaling coefficient

b: length-weight shape parameter

LenAge_cv: length-age coefficient of variability

Examples

data('YFT_221cell_observations_67_100_v2')
head(sim_67$obs$simulated_cpue_ll_jpn_86$data$obs)

#>      [,1]      [,2]                 
#> [1,] "r10-c4"  "84.406647306939036" 
#> [2,] "r11-c12" "2.0425311282486112" 
#> [3,] "r11-c14" "0.92323511903821065"
#> [4,] "r11-c15" "0.98871707919736485"
#> [5,] "r11-c16" "1.2425254930343868" 
#> [6,] "r11-c17" "38.91557805314195"