100 simulated YFT datasets at the 4 area scale

100 simulations of the YFT 1 area aggregated catch, cpue, length frequencies, and tagging data (along with biological data).

data('YFT_4area_observations_1-100')

Format

A data frame 'YFT_4area_observations_1_100.RData' with variables:

dat_4A_X

Simulated aggregated 4 area datasets for YFT; X=1:100

biol_dat

Biological data for YFT

Details

The key lists in YFT_SRD_1A_4$ (e.g., dat_1A_4$lencomp) include:

lencomp: (list) 1 area dataframe of aggregated length frequencies by length bin with columns corresponding to pseudo-year (Yr), season(Seas), fleet/survey index (FltSvy), gender (Gender), partition – length data come from all fish encountered (Part=0), number of lengths sampled (Nsamp), and the set of length bins (purse seine only)

catch: (list) 1 area dataframe of catch by fishery, including pseudo-year (year), and season (seas)

CPUE/cpu: (list) 1 area cpue with columns for pseudo-year (year), season (seas), cpue (cpu), standard deviation associated with lognormal error in cpue (cv), and survey index (index); note that CPUE is for longline only

tag_releases: 1 area tag release data with columns corresponding to generic tag release event number (tg), area of release (reg), pseudo-year (yr), season (season), gender (gender), age (age), and number of releases (nrel) adjusted for initial tag mortality

tag_recaps: 1 area tag recapture data with columns corresponding to generic tag release event number (tg), pseudo-year (yr), season (season), fleet index (fleet), and the number of recaptures (recaps)

The lists in biol_dat include:

M: age varying natural mortality age: first age to last age (pseudo-years) maturity: maturity at age ogive (pseudo-years) k: age varying growth coefficients L: length (cm) Linf: length infinity (cm) Lmin: length minimum (cm) a: length-weight scaling coefficient b: length-weight shape parameter

Examples

data('YFT_4area_observations_1_100')
head(dat_4A_1catch)

#> Error in head(dat_4A_1catch): object 'dat_4A_1catch' not found