START HERE (YFT): 4 area single representative dataset

A single representative dataset aggregated at the 4 area scale. The chosen dataset is simulation #4. List objects include aggregated catch, cpue, length frequencies, and tagging data (along with biological data).

data('YFT_SRD_4A_4')

Format

A data frame 'YFT_SRD_4A_4.RData' with variables:

dat_4A_X

Simulated aggregated 4 area datasets for YFT

biol_dat

Biological data for YFT

The key lists in YFT_SRD_4A_4$ (e.g., dat_4A_4$lencomp) include:

lencomp: (list) 1 area dataframe of aggregated length frequencies by length bin with columns corresponding to pseudo-year (Yr), season(Seas), fleet/survey index (FltSvy), sex (Gender), partition – length data come from all fish encountered (Part=0), number of lengths sampled (Nsamp), and the set of length bins (purse seine only)

catch: (list) 1 area dataframe of catch by fishery, including pseudo-year (year), and season (seas)

CPUE/cpu: (list) 1 area cpue with columns for pseudo-year (year), season (seas), cpue (cpu), standard deviation associated with lognormal error in cpue (cv), and survey index (index); note that CPUE is for longline only

tag_releases: 1 area tag release data with columns corresponding to generic tag release event number (tg), area of release (reg), pseudo-year (yr), season (season), sex (Gender), age (age), and number of releases (nrel) adjusted for initial tag mortality

tag_recaps: 1 area tag recapture data with columns corresponding to generic tag release event number (tg), pseudo-year (yr), season (season), fleet index (fleet), and the number of recaptures (recaps)

Explanations of list object items in the provided data files that analysts may find additionally useful.

Much of these follow, or closely follow, Stock Synthesis (version SS3.24Z) naming conventions. The name is the list identifier in each Rdata object (e.g., the name styr is found using dat_1A_4$styr).

styr Start year (pseudo-year)

endyr End year (pseudo-year)

nseas Number of seasons

spawn_seas Season when spawning occurs

Nfleet Number of fishing fleets

Nsurveys Number of surveys (fishery-dependent CPUE)

N_areas Number of spatial areas

fleetnames Specific fleet and survey index names

surveytiming Timing of fleet and surveys

areas Areas where each fleet or survey operate

units_of_catch Specifies units for catch (2=numbers of fish)

se_log_catch Standard error associated with catch

Ngenders Number of sexes

Nages Number of ages

lbin_vector_pop Vector of length bins in the population (cm)

lbin_vector Vector of length bins in the data (cm)

N_tag_groups Number of tag release events

N_recap_events Number of recapture events

mixing_latency_period Length of time units (psudo-years) necessary to allow for tag mixing

The lists in biol_dat include:

M: age varying natural mortality

age: first age to last age (pseudo-years)

maturity: maturity at age ogive (pseudo-years)

k: age varying growth coefficients

L: length (cm)

Linf: length infinity (cm)

Lmin: length minimum (cm)

a: length-weight scaling coefficient

b: length-weight shape parameter

Examples

data('YFT_SRD_4A_4')
head(dat_4A_4$catch)

#>   fishing_gi_1 fishing_gi_4 fishing_hd_1 fishing_ll_1 fishing_ll_2 fishing_ll_3
#> 1     13.71748     13.33279     5.710196            0            0            0
#> 2     13.71748     13.13353     5.710196            0            0            0
#> 3     13.71748     13.14095     5.710196            0            0            0
#> 4     13.71748     12.85522     5.710196            0            0            0
#> 5     15.73445     12.26299     5.713287            0            0            0
#> 6     15.73445     12.11916     5.713287            0            0            0
#>   fishing_ll_4 fishing_other_1 fishing_other_4 fishing_bb_1 fishing_ps_1
#> 1     0.000000      0.02576857        2.652977     92.61856            0
#> 2     0.000000      0.02576857        2.652977     92.61856            0
#> 3     0.000000      0.02576857        2.652977     92.61856            0
#> 4     0.000000      0.02576857        2.652977     92.61856            0
#> 5     0.000000      0.02938521        2.073448     93.03106            0
#> 6     4.612134      0.02938521        2.073448     93.03106            0
#>   fishing_ps_2 fishing_ps_4 fishing_trol_1 fishing_trol_2 fishing_trol_4 year
#> 1            0            0       5.138943       12.91881       45.93675    1
#> 2            0            0       5.138943       12.91881       45.93675    2
#> 3            0            0       5.138943       12.91881       45.93675    3
#> 4            0            0       5.138943       12.91881       45.93675    4
#> 5            0            0       5.700175       13.27055       43.62298    5
#> 6            0            0       5.700175       13.27055       43.62298    6
#>   seas
#> 1    1
#> 2    1
#> 3    1
#> 4    1
#> 5    1
#> 6    1

head(biol_dat)

#> $B0
#> [1] 97250600
#> 
#> $Maturity
#>  [1] 0.00 0.00 0.00 0.00 0.10 0.15 0.20 0.30 0.50 0.70 0.90 1.00 1.00 1.00 1.00
#> [16] 1.00 1.00 1.00 1.00 1.00 1.00 1.00 1.00 1.00 1.00 1.00 1.00 1.00
#> 
#> $M
#>  [1] 0.3358 0.2955 0.2552 0.2149 0.1746 0.1343 0.1343 0.1343 0.1343 0.1343
#> [11] 0.1410 0.1606 0.1780 0.1915 0.1994 0.2009 0.1962 0.1865 0.1743 0.1623
#> [21] 0.1522 0.1449 0.1401 0.1373 0.1357 0.1349 0.1346 0.1344
#> 
#> $Linf
#> [1] 145
#> 
#> $Lmin
#> [1] 22
#> 
#> $ages
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25
#> [26] 26 27 28
#>